requirements.txt
pygame>=2.1.0
numpy>=1.21.1
Pillow>=8.3.1
